clean words.
a) It is puzzled for novice likes me to understand what a 'project' is.
b) In description column, deleting makes concept clean, clear, and easy for novice to understand. Same words are said in "Changing config settings" and "How settings are applied". Please do not stuff lots in introduction section.