C32172: Removing <pre> tag that avoid content to be localizable
Hello, @zhili1208,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
"The writer used a functionality who blocks the content to be parsed for localization. 
If the intention is to have that string as localizable content, \<pre\> tag needs to be removed."

Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.