Update install-and-use-a-package-in-visual-studio.md
I got an error when trying to run this that stated: "Formatting" is an ambiguous reference between "Newtonsoft.Json.Formatting" and "System.Xml.Formatting", so I had to insert the "Newtonsoft.Json." before the "Formatting" myself. Assuming this could be an issue for other users, we should include this text here by default to make the tutorial as smooth as possible.