Update docs/create-packages/Sign-a-Package.md
Starting with .NET 9 and NuGet.exe 6.12, NU3043 warning is raised when a SHA-1 certificate fingerprint is passed to the sign commands. SHA-1 is considered insecure and should no longer be used.

This warning is promoted to an error in the .NET 10 SDK, and will be promoted to an error in NuGet.exe around .NET 10's release.

Co-authored-by: Kartheek Penagamuri <52756182+kartheekp-ms@users.noreply.github.com>